Add container testing + fix breakage (#22)
When we bumped the base Rust image, we changed how the PATH
is handled. This would cause rust-analyzer from /usr/local/cargo/bin
to be found first. This was not installed by default, so it
would trigger an indexing error when trying to use 'scip-rust index'.

This patch changes the Dockerfile to use the official rust-analyzer
component (previously, it wasn't bundled). It also adds a test for
the container, so that we don't accidentally regress functionality.